backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.7.0

# LlamaIndex core
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-fastembed>=0.1.0
tree-sitter-languages>=1.8.0  # For AST-aware code chunking

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# GitHub integration
PyGithub>=2.1.0
requests>=2.31.0

# Utilities
tiktoken>=0.5.0
websockets>=13.0
numpy>=1.21.0
